bug: if array of postingPreviews is size 0, doesn't reload properly (fixed) & code clean up
